Added changes to terminate compilation, on certain conditions #11
Changes are added to terminate the compilation, when different
variants are selected consecutively for compilation. As same
source file are being used during both CPP and C compilation
there are object files created from previous compilation
which hinders the process.This change makes sure, that the
object files are removed and compilation is smooth.